search_images/requirements.txt
regex
transformers
ftfy
flask
requests
gunicorn==19.3.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-aiplatform==1.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.34.2
google-cloud-core==1.7.2
google-cloud-logging==3.0.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
huggingface-hub==0.4.0
flask-cors
direct-redis
flask-healthz
streamlit==0.88
watchdog==2.1.6